% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chebcoefs}
\alias{chebcoefs}
\title{Function finding chebyshev coefficients}
\usage{
chebcoefs(fy, degree = 5)
}
\arguments{
\item{fy}{a square matrix with element (i,j) corresponding to f(x_j, y_i)}

\item{degree}{the degree of the cheybyshev polynomial}
}
\value{
a matrix of coefficients a_{i,j}
}
\description{
Function finding chebyshev coefficients
}
