% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chebeval}
\alias{chebeval}
\title{Function for evaluating chebyshev polynomials at nodes}
\usage{
chebeval(z, degree = 5)
}
\arguments{
\item{z}{a vector of nodes length i}

\item{degree}{the degree j of the cheybyshev polynomial}
}
\value{
a matrix with element (i,j) = T_j(z_i)
}
\description{
Function for evaluating chebyshev polynomials at nodes
}
