% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chebnodes}
\alias{chebnodes}
\title{Function for creating chebyshev nodes in 1 dimension}
\usage{
chebnodes(m)
}
\arguments{
\item{m}{a double - the number of nodes to create}
}
\value{
a vector of m nodes on [-1,1]
}
\description{
Function for creating chebyshev nodes in 1 dimension
}
