% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chebpred}
\alias{chebpred}
\title{Function for interpolating using chebyshev coefficients}
\usage{
chebpred(x0, y0, coefmat, lb, ub)
}
\arguments{
\item{x0}{x coordinate for interpolation}

\item{y0}{y coordinate for interpolation}

\item{coefmat}{matrix of coefficients from chebcoefs}

\item{lb}{a vector of 2 lower bounds in each dimension}

\item{ub}{a vector of 2 upper bounds in each dimension}
}
\value{
approximation for f(x,y)
}
\description{
Function for interpolating using chebyshev coefficients
}
