% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chebstatespace}
\alias{chebstatespace}
\title{Function for creating a 2d state space of nodes}
\usage{
chebstatespace(m, lb, ub)
}
\arguments{
\item{m}{a double - the number of nodes to create in each dimension}

\item{lb}{a vector of 2 lower bounds in each dimension}

\item{ub}{a vector of 2 upper bounds in each dimension}
}
\value{
a dataframe with all combinations of nodes to pass for function evaluation using mapply or similar
}
\description{
Function for creating a 2d state space of nodes
}
