% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{complexr}
\alias{complexr}
\title{Function for simplical interpolation in 3}
\usage{
complexr(x0, x, vfx)
}
\arguments{
\item{x0}{coordinates of point for interpolation}

\item{x}{matrix with x(i, j) the coordinate of point i in dimension j - must be a crossing}

\item{vfx}{function values at each x}
}
\value{
interpolated value at x0
}
\description{
Function for simplical interpolation in 3
}
