% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simplr}
\alias{simplr}
\title{Function for bilinear and simplical interpolation in 2D}
\usage{
simplr(x0, y0, x, y, vfx, method = "bilinear")
}
\arguments{
\item{x0}{coordinate for interpolation}

\item{y0}{coordinate for interpolation}

\item{x}{vector of x coordinates with known function values}

\item{y}{vector of y coordinates with known function values}

\item{vfx}{function values at each (x,y). Note x and y must form a crossing - must have function values at each possible combination}

\item{method}{string either "bilinear" or "simplical"}
}
\value{
interpolated value at x0 and y0
}
\description{
Function for bilinear and simplical interpolation in 2D
}
